backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.31
pydantic==2.7.0
pydantic-settings==2.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
openai>=1.30.0
